+++ frontend/lib/src/components/elements/DeckGlJsonChart/DeckGlJsonChart.test.tsx
@@ -14,18 +14,26 @@
  * limitations under the License.
  */
 
-import React from "react"
+import React, { FC } from "react"
 
-import { screen, waitFor } from "@testing-library/react"
+import { PickingInfo } from "@deck.gl/core"
+import { act, screen } from "@testing-library/react"
 import { userEvent } from "@testing-library/user-event"
+import JSON5 from "json5"
 
 import { DeckGlJsonChart as DeckGlJsonChartProto } from "@streamlit/protobuf"
 
-import { render } from "~lib/components/shared/ElementFullscreen/testUtils"
+import { ElementFullscreenContext } from "~lib/components/shared/ElementFullscreen/ElementFullscreenContext"
+import {
+  render,
+  renderHook,
+} from "~lib/components/shared/ElementFullscreen/testUtils"
+import { useRequiredContext } from "~lib/hooks/useRequiredContext"
+import { mockTheme } from "~lib/mocks/mockTheme"
 import { WidgetStateManager } from "~lib/WidgetStateManager"
 
-import { DeckGlJsonChart } from "./DeckGlJsonChart"
 import type { DeckGLProps } from "./types"
+import { useDeckGl, UseDeckGlProps } from "./useDeckGl"
 
 const mockInitialViewState = {
   bearing: -27.36,
@@ -38,11 +46,9 @@ const mockInitialViewState = {
   zoom: 6,
 }
 
-const mockHasLightBackgroundColor = vi.fn(() => false)
-
 vi.mock("~lib/theme", async () => ({
-  ...(await vi.importActual<typeof import("~lib/theme")>("~lib/theme")),
-  hasLightBackgroundColor: () => mockHasLightBackgroundColor(),
+  ...(await vi.importActual("~lib/theme")),
+  hasLightBackgroundColor: vi.fn(() => false),
 }))
 
 const getProps = (
@@ -87,104 +93,320 @@ const getProps = (
   }
 }
 
-describe("DeckGlJsonChart", () => {
-  it("should render with correct test id and className", () => {
-    const props = getProps()
-    render(<DeckGlJsonChart {...props} />)
-    const element = screen.getByTestId("stDeckGlJsonChart")
-    expect(element).toBeVisible()
-    expect(element).toHaveClass("stDeckGlJsonChart")
+const getUseDeckGlProps = (
+  elementProps: Partial<DeckGlJsonChartProto> = {},
+  initialViewStateProps: Record<string, unknown> = {}
+): UseDeckGlProps => {
+  return {
+    ...getProps(elementProps, initialViewStateProps),
+    isLightTheme: false,
+    theme: mockTheme.emotion,
+  }
+}
+
+describe("#useDeckGl", () => {
+  it("should merge client and server changes in viewState", () => {
+    const initialProps = getUseDeckGlProps()
+
+    const {
+      result: { current },
+      rerender,
+    } = renderHook(props => useDeckGl(props), {
+      initialProps,
+    })
+
+    expect(current.viewState).toEqual(mockInitialViewState)
+
+    rerender({
+      ...initialProps,
+      element: getUseDeckGlProps({}, { zoom: 8 }).element,
+    })
+
+    // should match original mockInitialViewState
+    expect(current.viewState).toEqual({ ...mockInitialViewState, zoom: 6 })
   })
 
-  it.each([
-    {
-      description: "no active selection",
-      setupSelection: false,
-      disabled: false,
-    },
-    {
-      description: "disabled state with selection",
-      setupSelection: true,
-      disabled: true,
-    },
-  ])(
-    "should not render clear selection button when $description",
-    ({ setupSelection, disabled }) => {
-      const props = getProps({
-        selectionMode: [DeckGlJsonChartProto.SelectionMode.SINGLE_OBJECT],
+  describe("createTooltip", () => {
+    it("should return null if info is null", () => {
+      const {
+        result: { current },
+      } = renderHook(props => useDeckGl(props), {
+        initialProps: getUseDeckGlProps(),
       })
 
-      if (setupSelection) {
-        props.widgetMgr.setStringValue(
-          props.element,
-          JSON.stringify({
+      expect(current.createTooltip(null)).toBe(null)
+    })
+
+    it("should return null if info.object is undefined", () => {
+      const {
+        result: { current },
+      } = renderHook(props => useDeckGl(props), {
+        initialProps: getUseDeckGlProps(),
+      })
+
+      expect(current.createTooltip({} as PickingInfo)).toBe(null)
+    })
+
+    it("should return null if element.tooltip is undefined", () => {
+      const {
+        result: { current },
+      } = renderHook(props => useDeckGl(props), {
+        initialProps: getUseDeckGlProps({ tooltip: undefined }),
+      })
+
+      expect(current.createTooltip({ object: {} } as PickingInfo)).toBe(null)
+    })
+
+    it("should interpolate the html with the correct object", () => {
+      const {
+        result: { current },
+      } = renderHook(props => useDeckGl(props), {
+        initialProps: getUseDeckGlProps({
+          tooltip: JSON.stringify({
+            html: "<b>Elevation Value:</b> {elevationValue}",
+          }),
+        }),
+      })
+
+      const result = current.createTooltip({
+        object: { elevationValue: 10 },
+      } as PickingInfo)
+
+      if (result === null || typeof result !== "object") {
+        throw new Error("Expected result to be an object")
+      }
+
+      expect(result.html).toBe("<b>Elevation Value:</b> 10")
+    })
+
+    it("should interpolate the html from object with a properties field", () => {
+      const {
+        result: { current },
+      } = renderHook(props => useDeckGl(props), {
+        initialProps: getUseDeckGlProps({
+          tooltip: JSON.stringify({
+            html: "<b>Elevation Value:</b> {elevationValue}",
+          }),
+        }),
+      })
+
+      const result = current.createTooltip({
+        object: { properties: { elevationValue: 10 } },
+      } as PickingInfo)
+
+      if (result === null || typeof result !== "object") {
+        throw new Error("Expected result to be an object")
+      }
+
+      expect(result.html).toBe("<b>Elevation Value:</b> 10")
+    })
+
+    it("should return the tooltip unchanged when object does have an expected schema", () => {
+      const {
+        result: { current },
+      } = renderHook(props => useDeckGl(props), {
+        initialProps: getUseDeckGlProps({
+          tooltip: JSON.stringify({
+            html: "<b>Elevation Value:</b> {elevationValue}",
+          }),
+        }),
+      })
+
+      const result = current.createTooltip({
+        object: { unexpectedSchema: { elevationValue: 10 } },
+      } as PickingInfo)
+
+      if (result === null || typeof result !== "object") {
+        throw new Error("Expected result to be an object")
+      }
+
+      expect(result.html).toBe("<b>Elevation Value:</b> {elevationValue}")
+    })
+
+    it("should interpolate the html with the an empty string", () => {
+      const {
+        result: { current },
+      } = renderHook(props => useDeckGl(props), {
+        initialProps: getUseDeckGlProps({ tooltip: "" }),
+      })
+
+      const result = current.createTooltip({
+        object: { elevationValue: 10 },
+      } as PickingInfo)
+
+      expect(result).toBe(null)
+    })
+  })
+
+  describe("deck", () => {
+    const newJson = {
+      initialViewState: mockInitialViewState,
+      mapStyle: "mapbox://styles/mapbox/light-v9",
+    }
+
+    const mockJsonParse = vi.fn().mockReturnValue(newJson)
+
+    beforeEach(() => {
+      JSON5.parse = mockJsonParse
+    })
+
+    afterEach(() => {
+      mockJsonParse.mockClear()
+    })
+
+    const testCases: {
+      description: string
+      newProps: Partial<UseDeckGlProps>
+    }[] = [
+      {
+        description: "should call JSON5.parse when the json is different",
+        newProps: getUseDeckGlProps(undefined, { zoom: 19 }),
+      },
+      {
+        description: "should call JSON5.parse when theme state changes",
+        newProps: { isLightTheme: true },
+      },
+    ]
+
+    it.each(testCases)("$description", ({ newProps }) => {
+      const initialProps = getUseDeckGlProps()
+      const { rerender } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+
+      expect(JSON5.parse).toHaveBeenCalledTimes(1)
+
+      rerender({ ...initialProps, ...newProps })
+
+      expect(JSON5.parse).toHaveBeenCalledTimes(2)
+    })
+
+    it("should call JSON5.parse when isFullScreen changes", async () => {
+      const user = userEvent.setup()
+      const MyComponent: FC<UseDeckGlProps> = props => {
+        useDeckGl(props)
+        const { expand } = useRequiredContext(ElementFullscreenContext)
+
+        return (
+          <button type="button" onClick={expand}>
+            Expand
+          </button>
+        )
+      }
+
+      render(<MyComponent {...getUseDeckGlProps()} />)
+
+      expect(JSON5.parse).toHaveBeenCalledTimes(1)
+
+      await user.click(screen.getByText("Expand"))
+
+      expect(JSON5.parse).toHaveBeenCalledTimes(2)
+    })
+  })
+
+  describe("selectionMode", () => {
+    it("should be undefined when allSelectionModes is empty", () => {
+      const initialProps = getUseDeckGlProps({ selectionMode: [] })
+      const { result } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+      expect(result.current.selectionMode).toBeUndefined()
+    })
+
+    it("should be defined when allSelectionModes has single object select", () => {
+      const initialProps = getUseDeckGlProps({
+        selectionMode: [DeckGlJsonChartProto.SelectionMode.SINGLE_OBJECT],
+      })
+      const { result } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+      expect(result.current.selectionMode).toBe(
+        DeckGlJsonChartProto.SelectionMode.SINGLE_OBJECT
+      )
+    })
+
+    it("should be defined when allSelectionModes has multi object select", () => {
+      const initialProps = getUseDeckGlProps({
+        selectionMode: [DeckGlJsonChartProto.SelectionMode.MULTI_OBJECT],
+      })
+      const { result } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+      expect(result.current.selectionMode).toBe(
+        DeckGlJsonChartProto.SelectionMode.MULTI_OBJECT
+      )
+    })
+
+    it("should return the first selection mode given, if multiple are given", () => {
+      const initialProps = getUseDeckGlProps({
+        selectionMode: [
+          DeckGlJsonChartProto.SelectionMode.MULTI_OBJECT,
+          DeckGlJsonChartProto.SelectionMode.SINGLE_OBJECT,
+        ],
+      })
+      const { result } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+      expect(result.current.selectionMode).toBe(
+        DeckGlJsonChartProto.SelectionMode.MULTI_OBJECT
+      )
+    })
+  })
+
+  describe("isSelectionModeActivated", () => {
+    it("should activate selection mode when selectionMode is defined", () => {
+      const initialProps = getUseDeckGlProps({
+        selectionMode: [DeckGlJsonChartProto.SelectionMode.SINGLE_OBJECT],
+      })
+      const { result } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+      expect(result.current.isSelectionModeActivated).toBe(true)
+    })
+
+    it("should not activate selection mode when selectionMode is undefined", () => {
+      const initialProps = getUseDeckGlProps({ selectionMode: [] })
+      const { result } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+      expect(result.current.isSelectionModeActivated).toBe(false)
+    })
+  })
+
+  describe("hasActiveSelection", () => {
+    it("should be false when selection is empty", () => {
+      const initialProps = getUseDeckGlProps({
+        selectionMode: [DeckGlJsonChartProto.SelectionMode.SINGLE_OBJECT],
+      })
+      const { result } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+      expect(result.current.hasActiveSelection).toBe(false)
+    })
+
+    it("should be true when selection is not empty", () => {
+      const initialProps = getUseDeckGlProps({
+        selectionMode: [DeckGlJsonChartProto.SelectionMode.SINGLE_OBJECT],
+      })
+      const { result, rerender } = renderHook(props => useDeckGl(props), {
+        initialProps,
+      })
+
+      act(() => {
+        result.current.setSelection({
+          fromUi: true,
+          value: {
             selection: {
               indices: { "0533490f-fcf9-4dc0-8c94-ae4fbd42eb6f": [0] },
               objects: { "0533490f-fcf9-4dc0-8c94-ae4fbd42eb6f": [{}] },
             },
-          }),
-          { fromUi: true },
-          props.fragmentId
-        )
-      }
-
-      render(<DeckGlJsonChart {...props} disabled={disabled} />)
-
-      expect(
-        screen.queryByLabelText("Clear selection")
-      ).not.toBeInTheDocument()
-    }
-  )
-
-  it.each([
-    {
-      description: "with mapbox token",
-      elementProps: { mapboxToken: "pk.test-token" },
-    },
-    {
-      description: "with default mapbox style",
-      elementProps: {},
-    },
-    {
-      description: "without selection mode",
-      elementProps: { selectionMode: [] },
-    },
-    {
-      description: "with single object selection mode",
-      elementProps: {
-        selectionMode: [DeckGlJsonChartProto.SelectionMode.SINGLE_OBJECT],
-      },
-    },
-  ])("should render correctly $description", ({ elementProps }) => {
-    const props = getProps(elementProps)
-    render(<DeckGlJsonChart {...props} />)
-    expect(screen.getByTestId("stDeckGlJsonChart")).toBeVisible()
-  })
-
-  describe("fullscreen mode", () => {
-    it("should render expand button by default", async () => {
-      const props = getProps()
-      render(<DeckGlJsonChart {...props} />)
-      const chart = screen.getByTestId("stDeckGlJsonChart")
-      await userEvent.hover(chart)
-
-      await waitFor(() => {
-        expect(screen.getByLabelText("Fullscreen")).toBeVisible()
+          },
+        })
       })
-    })
 
-    it("should not render fullscreen button when disableFullscreenMode is true", async () => {
-      const props = getProps()
-      render(<DeckGlJsonChart {...props} disableFullscreenMode />)
-      const chart = screen.getByTestId("stDeckGlJsonChart")
-      await userEvent.hover(chart)
+      rerender(initialProps)
 
-      // We use a hardcoded timeout here because we're testing a negative assertion
-      // (that something does NOT appear). Unlike positive assertions where we can
-      // wait for an element to appear, there's no reliable way to "wait for something
-      // to not appear" - we need to give sufficient time for it to potentially render.
-      await new Promise(resolve => setTimeout(resolve, 100))
-      expect(screen.queryByLabelText("Fullscreen")).not.toBeInTheDocument()
+      expect(result.current.hasActiveSelection).toBe(true)
     })
   })
 })
