+++ frontend/lib/src/components/elements/DeckGlJsonChart/utils/colors.test.ts
@@ -14,11 +14,7 @@
  * limitations under the License.
  */
 
-import {
-  getContextualFillColor,
-  LAYER_TYPE_TO_FILL_FUNCTION,
-  SerializedColorArray,
-} from "./colors"
+import { getContextualFillColor, SerializedColorArray } from "./colors"
 
 describe("#getContextualFillColor", () => {
   const object = { count: 10 }
@@ -168,174 +164,5 @@ describe("#getContextualFillColor", () => {
 
   it.each(testCases)("%s", (_description, args, expected) => {
     expect(getContextualFillColor(args)).toEqual(expected)
-  })
-
-  describe("fallback colors", () => {
-    it("should return selected fallback color when originalFillFunction returns non-array value and item is selected", () => {
-      const result = getContextualFillColor({
-        isSelected: true,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: () => 123, // Returns a number, not an array
-        selectedColor,
-        unselectedColor,
-      })
-      expect(result).toEqual(selectedColor)
-    })
-
-    it("should return unselected fallback color when originalFillFunction returns non-array value and item is not selected", () => {
-      const result = getContextualFillColor({
-        isSelected: false,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: () => 123, // Returns a number, not an array
-        selectedColor,
-        unselectedColor,
-      })
-      expect(result).toEqual(unselectedColor)
-    })
-
-    it("should return selected fallback color when originalFillFunction is undefined and item is selected", () => {
-      const result = getContextualFillColor({
-        isSelected: true,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: undefined,
-        selectedColor,
-        unselectedColor,
-      })
-      expect(result).toEqual(selectedColor)
-    })
-
-    it("should return unselected fallback color when originalFillFunction is undefined and item is not selected", () => {
-      const result = getContextualFillColor({
-        isSelected: false,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: undefined,
-        selectedColor,
-        unselectedColor,
-      })
-      expect(result).toEqual(unselectedColor)
-    })
-  })
-
-  describe("custom opacity values", () => {
-    it("should use custom selectedOpacity when provided", () => {
-      const result = getContextualFillColor({
-        isSelected: true,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: () => [100, 100, 100, 50],
-        selectedColor,
-        unselectedColor,
-        selectedOpacity: 200,
-      })
-      // selectedOpacity is 200, original is 50, so max(50, 200) = 200
-      expect(result).toEqual([100, 100, 100, 200])
-    })
-
-    it("should use custom unselectedOpacity when provided", () => {
-      const result = getContextualFillColor({
-        isSelected: false,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: () => [100, 100, 100, 255],
-        selectedColor,
-        unselectedColor,
-        unselectedOpacity: 50,
-      })
-      // unselectedOpacity is 50, original is 255, so min(255, 50) = 50
-      expect(result).toEqual([100, 100, 100, 50])
-    })
-
-    it("should respect original lower opacity when unselected", () => {
-      const result = getContextualFillColor({
-        isSelected: false,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: () => [100, 100, 100, 30],
-        selectedColor,
-        unselectedColor,
-        unselectedOpacity: 100,
-      })
-      // unselectedOpacity is 100, original is 30, so min(30, 100) = 30
-      expect(result).toEqual([100, 100, 100, 30])
-    })
-  })
-
-  describe("two-element color arrays", () => {
-    it("should handle two-element color array", () => {
-      const result = getContextualFillColor({
-        isSelected: false,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: () => [255, 128],
-        selectedColor,
-        unselectedColor,
-      })
-      // [255, 128] -> [255, 128, 0, 102] (with default unselected opacity)
-      expect(result).toEqual([255, 128, 0, 102])
-    })
-
-    it("should handle three-element color array", () => {
-      const result = getContextualFillColor({
-        isSelected: false,
-        object,
-        objectInfo: { index: 0 },
-        originalFillFunction: () => [255, 128, 64],
-        selectedColor,
-        unselectedColor,
-      })
-      // [255, 128, 64] -> [255, 128, 64, 102] (with default unselected opacity)
-      expect(result).toEqual([255, 128, 64, 102])
-    })
-  })
-})
-
-describe("#LAYER_TYPE_TO_FILL_FUNCTION", () => {
-  it("should have all values as arrays of strings", () => {
-    // Verify the structure of the mapping: all values should be arrays of strings
-    Object.entries(LAYER_TYPE_TO_FILL_FUNCTION).forEach(([_key, value]) => {
-      expect(Array.isArray(value)).toBe(true)
-      value.forEach(fnName => {
-        expect(typeof fnName).toBe("string")
-      })
-    })
-  })
-
-  it("should map ScatterplotLayer to multiple fill functions", () => {
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.ScatterplotLayer).toContain(
-      "getFillColor"
-    )
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.ScatterplotLayer).toContain("getColor")
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.ScatterplotLayer).toContain(
-      "getLineColor"
-    )
-  })
-
-  it("should map ArcLayer to source and target color functions", () => {
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.ArcLayer).toContain("getTargetColor")
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.ArcLayer).toContain("getSourceColor")
-  })
-
-  it("should map GeoJsonLayer to getFillColor", () => {
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.GeoJsonLayer).toContain("getFillColor")
-  })
-
-  it("should map IconLayer to getColor", () => {
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.IconLayer).toContain("getColor")
-  })
-
-  it("should map LineLayer to getColor", () => {
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.LineLayer).toContain("getColor")
-  })
-
-  it("should map PathLayer to getColor", () => {
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.PathLayer).toContain("getColor")
-  })
-
-  it("should map TextLayer to getColor", () => {
-    expect(LAYER_TYPE_TO_FILL_FUNCTION.TextLayer).toContain("getColor")
   })
 })
